cleaned up code and moved functions to utils file
